--- backend/db.js
@@ -1,3 +1,0 @@
-module.exports = {
-    DB: 'mongodb://localhost:27017/inventory'
-}
