+++ client/src/components/Inventory/Inventory.css
@@ -33,7 +33,7 @@
 .form-control2 {
     margin: 19px auto;
     display: block;
-    width: 25%;
+    width: 70%;
     height: 34px;
     padding: 6px 12px;
     font-size: 14px;
@@ -50,4 +50,17 @@
     transition: border-color ease-in-out .15s,-webkit-box-shadow ease-in-out .15s;
     transition: border-color ease-in-out .15s,box-shadow ease-in-out .15s;
     transition: border-color ease-in-out .15s,box-shadow ease-in-out .15s,-webkit-box-shadow ease-in-out .15s;
+}
+
+.btn-primary {
+    color: #fff;
+    background-color: #333;
+    border-color: #333;
+    margin: 10px 0px 0px 50px;
+}
+.btn-primary:hover {
+    color: grey;
+    background-color: black;
+    border-color: black;
+    margin: 10px 0 0 50px;
 }
