+++ client/src/components/Register.css
@@ -1,50 +1,3 @@
-
-#body{
-    /* background-color; */
-    font-size: 14px;
-}
-p{
-    color:yellow;
-    font-weight: bold;
-    font-size: 20px;
-}
-label {
-    display: inline-block;
-    max-width: 100%;
-    font-weight: 100;
-    font-size: medium;
-    margin: 25px 0px 0px 0px;
-}
-input, optgroup, select, textarea {
-    margin: 20px 0px 0 52px;
-    font: inherit;
-    color: inherit;
-    width: 77%;
-}
-.production-menu{
-    width:80%;
-    height: 450px;
-    background-color:slategray;
-    margin: auto;
-    text-align: left;
-    padding-left:10%;
-}
-.user-input-p{
-    background-color:rosybrown;
-    width:50%;
-    padding-left:20%;
-}
-#label{
-    background-color:blue;
-    color:white;
-    margin-bottom:0;
-    margin-top:10px;
-}
-.col-sm-3 {
-    width: 22.9%;
-    text-align: right;
-}
-
 .btn-primary {
     color: grey;
     background-color: black;
@@ -57,6 +10,7 @@ input, optgroup, select, textarea {
     border-color: black;
     margin-right: 0px;
 }
+
 .form-control {
     margin-bottom: 15px;
     display: block;
