--- client/src/components/User/NewUser.css
@@ -1,102 +1,0 @@
-
-#body{
-    /* background-color; */
-    font-size: 14px;
-}
-label {
-    display: inline-block;
-    max-width: 100%;
-    font-weight: 100;
-    font-size: medium;
-    margin: 25px 0px 0px 0px;
-}
-input, optgroup, select, textarea {
-    margin: 20px 0px 0 52px;
-    font: inherit;
-    color: inherit;
-}
-.navbar-nav {
-    float: left;
-    margin: 0;
-    background-color: black;
-}
-
-.btn-primary {
-    color: #fcf8e3;
-    background-color: black;
-    border-color: black;
-    margin: 0px 0px 0px 145px;
-}
-.btn-primary:hover {
-    color: grey;
-    background-color: black;
-    border-color: black;
-}
-.col-sm-10 {
-    width: 81.333333%;
-}
-.col-sm-2 {
-    width: 18.666667%;
-}
-.user-menu{
-    width:80%;
-    height: 400px;
-    margin: auto;
-    text-align: left;
-    padding-left:10%;
-}
-.user-input-u{
-    width:50%;
-    margin:auto;
-}
-#selectbtn{
-    color:black;
-    width:150px;
-}
-#current-date{
-    width:30%;
-}
-.savebtn{
-    width:200px;
-    color:black;
-}
-
-#user_id, #user_password, #user_email{
-    width:60%;
-}
-.usersavebtn {
-    width: 174px;
-    background-color: white;
-    color: black;
-    font-size: 15px;
-    font-weight: normal;
-    margin: 9px 0 0 356px;
-}
-#email{
-    margin: 0px 0px 0px 0px;
-}
-.form-control {
-    margin-bottom: 15px;
-    display: block;
-    width: 90%;
-    height: 34px;
-    padding: 6px 12px;
-    font-size: 14px;
-    line-height: 1.42857143;
-    color: #555;
-    background-color: #fff;
-    background-image: none;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    -webkit-box-shadow: inset 0 1px 1px rgba(0,0,0,.075);
-    box-shadow: inset 0 1px 1px rgba(0,0,0,.075);
-    -webkit-transition: border-color ease-in-out .15s,-webkit-box-shadow ease-in-out .15s;
-    -o-transition: border-color ease-in-out .15s,box-shadow ease-in-out .15s;
-    transition: border-color ease-in-out .15s,box-shadow ease-in-out .15s;
-}
-.modal-footer {
-    padding: 15px;
-    text-align: right;
-    border-top: 1px solid #e5e5e5;
-    margin-top: 28rem;
-}
